microapps-publish v2 - muti-command
