+++ .github/workflows/main-build.yml
@@ -35,11 +35,10 @@ jobs:
           git config user.name "Automation"
           git config user.email "github-actions@github.com"
       
-      - name: Fix Install for microapps-publish
+      - name: Fix Permissions
         run: |
           chown -R root ./
           chmod -R 777 ./
-          npm run fixinstall
 
       - name: Print NPM Cache Location
         run: npm config get cache
@@ -57,8 +56,8 @@ jobs:
         run: npm ci
 
       - name: Generate Projen Files
+        working-directory: packages/microapps-cdk/
         run: |
-          cd packages/microapps-cdk/
           npm run projen
 
       - name: Build All TypeScript
@@ -68,13 +67,13 @@ jobs:
         run: mv node_modules node_modules_hide
 
       - name: Install CDK Construct Deps
+        working-directory: packages/microapps-cdk/
         run: |
-          cd packages/microapps-cdk
           npm ci
 
       - name: Build CDK Construct
+        working-directory: packages/microapps-cdk/
         run: |
-          cd packages/microapps-cdk
           npm run build
 
       - name: Check for new commits
@@ -115,13 +114,6 @@ jobs:
           node-version: 16
           cache: 'npm'
           cache-dependency-path: '**/package-lock.json'
-
-      - name: Fix Install for microapps-publish
-        run: |
-          npm run fixinstall
-
-      - name: Install CDK
-        run: npm i -g aws-cdk ts-node projen
 
       - name: Check for new commits
         id: git_remote
@@ -194,13 +186,13 @@ jobs:
         run: npm set //registry.npmjs.org/:_authToken ${{ secrets.NPMJSORG_PUBLISH_TOKEN }}
 
       - name: Release Publish Tool - NPM
+        working-directory: packages/microapps-publish
         run: |
-          cd packages/microapps-publish
           npm publish --dry-run
 
       - name: Release DataLib - NPM
+        working-directory: packages/microapps-datalib
         run: |
-          cd packages/microapps-datalib
           npm publish --dry-run
 
     container:
