+++ .github/workflows/release.yml
@@ -36,11 +36,10 @@ jobs:
           git config user.name "Automation"
           git config user.email "github-actions@github.com"
       
-      - name: Fix Install for microapps-publish
+      - name: Fix Permissions
         run: |
           chown -R root ./
           chmod -R 777 ./
-          npm run fixinstall
 
       - name: Print NPM Cache Location
         run: npm config get cache
@@ -58,8 +57,8 @@ jobs:
         run: npm ci
 
       - name: Generate Projen Files
+        working-directory: packages/microapps-cdk/
         run: |
-          cd packages/microapps-cdk/
           npm run projen
 
       - name: Build All TypeScript
@@ -69,13 +68,13 @@ jobs:
         run: mv node_modules node_modules_hide
 
       - name: Install CDK Construct Deps
+        working-directory: packages/microapps-cdk/
         run: |
-          cd packages/microapps-cdk
           npm ci
 
       - name: Build CDK Construct
+        working-directory: packages/microapps-cdk/
         run: |
-          cd packages/microapps-cdk
           npx projen release
 
       - name: Apply Version to Everything (Deployer / Datalib)
@@ -93,10 +92,10 @@ jobs:
           name: version-txt
           path: packages/microapps-cdk/dist/version.txt
 
-      - name: Create release
-        if: ${{ steps.git_remote.outputs.latest_commit == github.sha }}
+      - if: ${{ steps.git_remote.outputs.latest_commit == github.sha }}
+        name: Create release
+        working-directory: packages/microapps-cdk/
         run: |
-          cd packages/microapps-cdk
           gh release create v$(cat dist/version.txt) -F dist/changelog.md -t v$(cat dist/version.txt)
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
@@ -134,13 +133,6 @@ jobs:
           node-version: 16
           cache: 'npm'
           cache-dependency-path: '**/package-lock.json'
-
-      - name: Fix Install for microapps-publish
-        run: |
-          npm run fixinstall
-
-      - name: Install CDK
-        run: npm i -g aws-cdk ts-node projen
 
       - name: Check for new commits
         id: git_remote
@@ -226,8 +218,8 @@ jobs:
       #     NUGET_API_KEY: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Release CDK Construct - NuGet.org
+        working-directory: packages/microapps-cdk/
         run: |
-          cd packages/microapps-cdk
           npx -p jsii-release@latest jsii-release-nuget
         env:
           NUGET_API_KEY: ${{ secrets.NUGET_MICROAPPS }}
@@ -237,8 +229,8 @@ jobs:
       # Docs for jsii-release-maven
       # https://github.com/cdklabs/jsii-release#maven
       - name: Release CDK Construct - Maven
+        working-directory: packages/microapps-cdk/
         run: |
-          cd packages/microapps-cdk
           npx -p jsii-release@latest jsii-release-maven
         env:
           MAVEN_SERVER_ID: github
@@ -258,8 +250,8 @@ jobs:
       #     NPM_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Release CDK Construct - npmjs.org
+        working-directory: packages/microapps-cdk
         run: |
-          cd packages/microapps-cdk
           npx -p jsii-release@latest jsii-release-npm
         env:
           NPM_TOKEN: ${{ secrets.NPMJSORG_PUBLISH_TOKEN }} # jsii-release requires NPM_TOKEN as an env var
@@ -284,13 +276,13 @@ jobs:
         run: npm set //registry.npmjs.org/:_authToken ${{ secrets.NPMJSORG_PUBLISH_TOKEN }}
 
       - name: Release Publish Tool - NPM
+        working-directory: packages/microapps-publish/
         run: |
-          cd packages/microapps-publish
           npm publish
 
       - name: Release DataLib - NPM
+        working-directory: packages/microapps-datalib/
         run: |
-          cd packages/microapps-datalib
           npm publish
 
       # - name: Github package registry authentication
