+++ package.json
@@ -22,7 +22,6 @@
     "clean": "npm run clean:dist && npm run clean:modules",
     "clean:dist": "npm exec --workspaces -- npx rimraf dist && npx rimraf dist",
     "clean:modules": "npm exec --workspaces -- npx rimraf node_modules && npx rimraf node_modules",
-    "fixinstall": "mkdir -p packages/microapps-publish/dist/ && touch packages/microapps-publish/dist/index.js",
     "cloc": "cloc --exclude-dir=node_modules,dist,distb,cdk.out --exclude-ext=json .",
     "start:publish": "node packages/microapps-publish/dist/src/index.js",
     "start:publish:debug": "ts-node --inspect=5858 --debug-brk --ignore false packages/microapps-publish/src/index.ts",
@@ -45,6 +44,7 @@
     "@types/node": "^16.9.2",
     "@typescript-eslint/eslint-plugin": "^4.9.0",
     "@typescript-eslint/parser": "^4.9.0",
+    "aws-cdk": "^1.134.0",
     "cross-env": "^7.0.3",
     "esbuild": "^0.12.14",
     "eslint": "^7.15.0",
@@ -56,7 +56,7 @@
     "jest": "^26.6.3",
     "jest-dynalite": "^3.4.4",
     "prettier": "^2.2.1",
-    "projen": "^0.34.20",
+    "projen": "~0.34.20",
     "rimraf": "^3.0.2",
     "rollup": "^2.39.0",
     "rollup-plugin-node-externals": "^2.2.0",
